send mail odoo

+++ my_project/mybook/models/my_book.py
@@ -6,5 +6,5 @@ class BookInfo(models.Model):
     _rec_name = 'name_of_book'  # lấy tên theo record
 
     name_of_book = fields.Char(string="Tên sách")
-    is_expried_of_book = fields.Float(string="Thời hạn trả sách", default=2)
+    is_expried_of_book = fields.Integer(string="Thời hạn trả sách", default=2)
     number_of_money = fields.Float(string="Hệ số tính tiền trả muộn", default=5000)
